03_Implementacao/SemanticSearch/Bucket/requirements.txt
flask==2.0.1
numpy==1.21.0
faiss-cpu==1.7.2
requests==2.28.1
werkzeug==2.0.3
pyPDF2==3.0.1
python-docx==0.8.11
markdown==3.6
ollama==0.3.2
packaging==24.1
